+++ sketch-animated-koralgif/src/androidTest/kotlin/com/github/panpf/sketch/animated/koralgif/test/decode/GifDrawableDecoderTest.kt
@@ -25,6 +25,7 @@ import com.github.panpf.sketch.decode.internal.createInSampledTransformed
 import com.github.panpf.sketch.decode.supportKoralGif
 import com.github.panpf.sketch.drawable.AnimatableDrawable
 import com.github.panpf.sketch.drawable.GifDrawableWrapperDrawable
+import com.github.panpf.sketch.fetch.FetchResult
 import com.github.panpf.sketch.fetch.copy
 import com.github.panpf.sketch.getDrawableOrThrow
 import com.github.panpf.sketch.images.ResourceImages
@@ -33,6 +34,7 @@ import com.github.panpf.sketch.request.animatedTransformation
 import com.github.panpf.sketch.request.onAnimationEnd
 import com.github.panpf.sketch.request.onAnimationStart
 import com.github.panpf.sketch.request.repeatCount
+import com.github.panpf.sketch.source.AssetDataSource
 import com.github.panpf.sketch.source.DataFrom.LOCAL
 import com.github.panpf.sketch.test.singleton.sketch
 import com.github.panpf.sketch.test.utils.decode
@@ -136,6 +138,18 @@ class GifDrawableDecoderTest {
         // data error
         ImageRequest(context, ResourceImages.png.uri).let {
             val fetchResult = it.fetch(sketch)
+            factory.create(it.toRequestContext(sketch), fetchResult)
+        }.apply {
+            Assert.assertNull(this)
+        }
+
+        // Disguised, mimeType; data error
+        ImageRequest(context, ResourceImages.png.uri).let {
+            val fetchResult =
+                FetchResult(
+                    AssetDataSource(sketch, it, ResourceImages.png.resourceName),
+                    "image/gif"
+                )
             factory.create(it.toRequestContext(sketch), fetchResult)
         }.apply {
             Assert.assertNull(this)