+++ sketch-animated/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/animated/android/test/decode/GifMovieDecoderTest.kt
@@ -149,6 +149,18 @@ class GifMovieDecoderTest {
         }.apply {
             Assert.assertNull(this)
         }
+
+        // Disguised, mimeType; data error
+        ImageRequest(context, ResourceImages.png.uri).let {
+            val fetchResult =
+                FetchResult(
+                    AssetDataSource(sketch, it, ResourceImages.png.resourceName),
+                    "image/gif"
+                )
+            factory.create(it.toRequestContext(sketch), fetchResult)
+        }.apply {
+            Assert.assertNull(this)
+        }
     }
 
     @Test