+++ sketch-animated/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/animated/android/test/decode/WebpAnimatedDecoderTest.kt
@@ -159,6 +159,18 @@ class WebpAnimatedDecoderTest {
             )
             factory.create(it.toRequestContext(sketch), fetchResult)
         }.apply {
+            Assert.assertNotNull(this)
+        }
+
+        // Disguised, mimeType; data error
+        ImageRequest(context, ResourceImages.png.uri).let {
+            val fetchResult =
+                FetchResult(
+                    AssetDataSource(sketch, it, ResourceImages.png.resourceName),
+                    "image/webp"
+                )
+            factory.create(it.toRequestContext(sketch), fetchResult)
+        }.apply {
             Assert.assertNull(this)
         }
     }