+++ sketch-animated/src/androidMain/kotlin/com/github/panpf/sketch/decode/WebpAnimatedDecoder.kt
@@ -18,12 +18,11 @@ package com.github.panpf.sketch.decode
 import android.os.Build
 import androidx.annotation.RequiresApi
 import com.github.panpf.sketch.ComponentRegistry
-import com.github.panpf.sketch.source.DataSource
 import com.github.panpf.sketch.decode.internal.ImageDecoderAnimatedDecoder
-import com.github.panpf.sketch.decode.internal.ImageFormat
 import com.github.panpf.sketch.decode.internal.isAnimatedWebP
 import com.github.panpf.sketch.fetch.FetchResult
 import com.github.panpf.sketch.request.internal.RequestContext
+import com.github.panpf.sketch.source.DataSource
 
 /**
  * Adds animation webp support by AnimatedImageDrawable
@@ -62,11 +61,9 @@ class WebpAnimatedDecoder(
             if (
                 Build.VERSION.SDK_INT >= Build.VERSION_CODES.P
                 && !requestContext.request.disallowAnimatedImage
+                && fetchResult.headerBytes.isAnimatedWebP()
             ) {
-                val imageFormat = ImageFormat.parseMimeType(fetchResult.mimeType)
-                if ((imageFormat == null || imageFormat == ImageFormat.WEBP) && fetchResult.headerBytes.isAnimatedWebP()) {
-                    return WebpAnimatedDecoder(requestContext, dataSource)
-                }
+                return WebpAnimatedDecoder(requestContext, dataSource)
             }
             return null
         }